+++ apps/agentos/store/AgentDefinitions.mjs
@@ -5,8 +5,13 @@ import Store                from '../../../src/data/Store.mjs';
  * @class AgentOS.store.AgentDefinitions
  * @extends Neo.data.Store
  *
- * @summary Redacted Fleet Manager agent definition list for the settings pane. The seed row is a
+ * @summary Redacted Fleet Manager agent definition list — the shared fleet roster. The seed row is a
  * local bridge-state placeholder, not persisted registry data, and carries no credential bytes.
+ *
+ * Exposed as a **singleton** so the keeper-views that compose the cockpit bind one shared instance:
+ * `AgentOS.view.Accounts` writes redacted identities into it (after the Brain-side credential submit),
+ * and `AgentOS.view.FleetSettingsPanel`'s roster grid reads from it reactively. No credential bytes
+ * ever enter this Body-side store.
  */
 class AgentDefinitions extends Store {
     static config = {
@@ -15,6 +20,10 @@ class AgentDefinitions extends Store {
          * @protected
          */
         className: 'AgentOS.store.AgentDefinitions',
+        /**
+         * @member {Boolean} singleton=true
+         */
+        singleton: true,
         /**
          * @member {Object[]} data
          */
